+++ pcs_test/tier1/legacy/test_constraints.py
@@ -394,6 +394,18 @@ class ConstraintTest(unittest.TestCase, AssertPcsMixin):
         self.assertEqual(stdout, "")
         self.assertEqual(retval, 1)
 
+    def test_order_score(self):
+        self.fixture_resources()
+        self.assert_pcs_success(
+            "constraint order D1 then D2 score=123".split(),
+            stderr_full=(
+                "Deprecation Warning: option 'score' is deprecated and might "
+                "be removed in a future release, therefore it should not be used\n"
+                "Adding D1 D2 (score: 123) (Options: first-action=start "
+                "then-action=start)\n"
+            ),
+        )
+
     def test_order_too_many_resources(self):
         msg = (
             "Error: Multiple 'then's cannot be specified.\n"
@@ -1668,7 +1680,8 @@ class ConstraintTest(unittest.TestCase, AssertPcsMixin):
             (
                 f"Deprecation Warning: Value "
                 f"'{const.PCMK_ROLE_UNPROMOTED_LEGACY}' of option "
-                "role is deprecated and should not be used, use "
+                "role is deprecated and might be removed in a future release, "
+                "therefore it should not be used, use "
                 f"'{const.PCMK_ROLE_UNPROMOTED}' value instead\n"
             ),
         )
@@ -4170,7 +4183,8 @@ class TicketAdd(ConstraintBaseTest):
             f"constraint ticket add T {role} A loss-policy=fence".split(),
             (
                 f"Deprecation Warning: Value '{role}' of option role is "
-                f"deprecated and should not be used, use "
+                "deprecated and might be removed in a future release, "
+                "therefore it should not be used, use "
                 f"'{const.PCMK_ROLE_UNPROMOTED}' value instead\n"
                 "Duplicate constraints:\n"
                 f"  {const.PCMK_ROLE_UNPROMOTED} resource 'A' depends on ticket 'T' (id: ticket-T-A-{const.PCMK_ROLE_UNPROMOTED})\n"
@@ -4186,7 +4200,8 @@ class TicketAdd(ConstraintBaseTest):
             f"constraint ticket add T {role} A loss-policy=fence".split(),
             stderr_full=(
                 f"Deprecation Warning: Value '{role}' of option role is "
-                f"deprecated and should not be used, use "
+                "deprecated and might be removed in a future release, "
+                "therefore it should not be used, use "
                 f"'{const.PCMK_ROLE_PROMOTED}' value instead\n"
             ),
         )
@@ -4317,7 +4332,8 @@ class TicketShow(ConstraintBaseTest):
             f"constraint ticket add T {role} A loss-policy=fence".split(),
             stderr_full=(
                 f"Deprecation Warning: Value '{role}' of option role is "
-                f"deprecated and should not be used, use "
+                "deprecated and might be removed in a future release, "
+                "therefore it should not be used, use "
                 f"'{const.PCMK_ROLE_PROMOTED}' value instead\n"
             ),
         )