+++ pcs_test/tier1/legacy/test_stonith.py
@@ -1438,7 +1438,8 @@ class StonithTest(TestCase, AssertPcsMixin):
         self.assert_pcs_success(
             "stonith create S fence_apc ip=i login=l password=1234".split(),
             stderr_full=(
-                "Warning: stonith option 'login' is deprecated and should not be "
+                "Warning: stonith option 'login' is deprecated and might be "
+                "removed in a future release, therefore it should not be "
                 "used, use 'username' instead\n"
             ),
         )
@@ -1473,9 +1474,11 @@ class StonithTest(TestCase, AssertPcsMixin):
                 "password=1234",
             ),
             stderr_full=(
-                "Warning: stonith option 'ipaddr' is deprecated and should not be "
+                "Warning: stonith option 'ipaddr' is deprecated and might be "
+                "removed in a future release, therefore it should not be "
                 "used, use 'ip' instead\n"
-                "Warning: stonith option 'login' is deprecated and should not be "
+                "Warning: stonith option 'login' is deprecated and might be "
+                "removed in a future release, therefore it should not be "
                 "used, use 'username' instead\n"
             ),
         )
@@ -1570,7 +1573,8 @@ class StonithTest(TestCase, AssertPcsMixin):
         self.assert_pcs_fail(
             "stonith create test fence_apc ip=i username=u action=a".split(),
             (
-                "Error: stonith option 'action' is deprecated and should not be"
+                "Error: stonith option 'action' is deprecated and might be "
+                "removed in a future release, therefore it should not be"
                 " used, use 'pcmk_off_action', 'pcmk_reboot_action' instead,"
                 " use --force to override\n" + ERRORS_HAVE_OCCURRED
             ),
@@ -1579,7 +1583,8 @@ class StonithTest(TestCase, AssertPcsMixin):
         self.assert_pcs_success(
             "stonith create test fence_apc ip=i username=u action=a --force".split(),
             stderr_start=(
-                "Warning: stonith option 'action' is deprecated and should not be "
+                "Warning: stonith option 'action' is deprecated and might be "
+                "removed in a future release, therefore it should not be "
                 "used, use 'pcmk_off_action', 'pcmk_reboot_action' instead\n"
             ),
         )
@@ -1630,7 +1635,8 @@ class StonithTest(TestCase, AssertPcsMixin):
         self.assert_pcs_fail(
             "stonith update test action=a".split(),
             (
-                "Error: stonith option 'action' is deprecated and should not be"
+                "Error: stonith option 'action' is deprecated and might be "
+                "removed in a future release, therefore it should not be"
                 " used, use 'pcmk_off_action', 'pcmk_reboot_action' instead,"
                 " use --force to override\n"
             ),
@@ -1639,7 +1645,8 @@ class StonithTest(TestCase, AssertPcsMixin):
         self.assert_pcs_success(
             "stonith update test action=a --force".split(),
             stderr_start=(
-                "Warning: stonith option 'action' is deprecated and should not be "
+                "Warning: stonith option 'action' is deprecated and might be "
+                "removed in a future release, therefore it should not be "
                 "used, use 'pcmk_off_action', 'pcmk_reboot_action' instead\n"
             ),
         )
@@ -3422,9 +3429,11 @@ class StonithUpdate(ResourceTest):
             </resources>
             """,
             stderr_regexp=(
-                "Warning: stonith option 'login' is deprecated and should not "
+                "Warning: stonith option 'login' is deprecated and might be "
+                "removed in a future release, therefore it should not "
                 "be used, use 'username' instead\n"
-                "Warning: stonith option 'debug' is deprecated and should not "
+                "Warning: stonith option 'debug' is deprecated and might be "
+                "removed in a future release, therefore it should not "
                 "be used, use 'debug_file' instead\n"
                 + self.agent_secure_warning
             ),
@@ -3462,7 +3471,8 @@ class StonithUpdate(ResourceTest):
             </resources>
             """,
             stderr_regexp=(
-                "Warning: stonith option 'debug' is deprecated and should not "
+                "Warning: stonith option 'debug' is deprecated and might be "
+                "removed in a future release, therefore it should not "
                 "be used, use 'debug_file' instead\n"
                 + self.agent_secure_warning
             ),
@@ -3642,7 +3652,8 @@ class StonithUpdate(ResourceTest):
             </resources>
             """,
             stderr_regexp=(
-                "Warning: stonith option 'ipaddr' is deprecated and should not "
+                "Warning: stonith option 'ipaddr' is deprecated and might be "
+                "removed in a future release, therefore it should not "
                 "be used, use 'ip' instead\n" + self.agent_secure_warning
             ),
         )
@@ -3682,7 +3693,8 @@ class StonithUpdate(ResourceTest):
             </resources>
             """,
             stderr_regexp=(
-                "Warning: stonith option 'ipaddr' is deprecated and should not "
+                "Warning: stonith option 'ipaddr' is deprecated and might be "
+                "removed in a future release, therefore it should not "
                 "be used, use 'ip' instead\n" + self.agent_secure_warning
             ),
         )