+++ pcs_test/tier1/test_status.py
@@ -38,7 +38,8 @@ class StonithWarningTest(TestCase, AssertPcsMixin):
                 "--force"
             ).split(),
             stderr_start=(
-                "Warning: stonith option 'action' is deprecated and should not be "
+                "Warning: stonith option 'action' is deprecated and might be "
+                "removed in a future release, therefore it should not be "
                 "used, use 'pcmk_off_action', 'pcmk_reboot_action' instead\n"
             ),
         )